==> spectral_flow_ics/7_Polyfunctionality.Rmd <==
---
title: "T cell polyfunctionality"
author: "*Lisa Loksø Dietz*"
date: "*`r format(Sys.time(), '%Y-%m-%d')`*"
knit: (function(inputFile, encoding) { 
      out_dir = paste0(dirname(getwd()),"/3_output");
      rmarkdown::render(inputFile,
                        encoding=encoding, 
                        output_file=file.path(out_dir, '7_TcellPolyfunctionality.html')) })
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 4
    toc_collapsed: true
    code_folding: show
---

```{r Setup 1, include = FALSE}

knitr::opts_chunk$set(fig.width=10, fig.height=7, out.width = "70%",fig.align = 'center', eval = FALSE) 

```


## 0) Load source code

```{r Directories, message = FALSE, warning=FALSE}

rm(list = ls())
source("0_source.R")

```


## 1) Load data

```{r Load SCE , message = FALSE}

sce_CD4 = readRDS(file = glue("{dir_project}/2_pipeline/6b_clustering2/out/sce_CD4.rds"))

sce_CD8 = readRDS(file = glue("{dir_project}/2_pipeline/6b_clustering2/out/sce_CD8.rds"))

```


## 2) Polyfunctionality

### 2.1) Add cytokine positivity

```{r}

#Cytokines
boolean_matrix_cytokines = data.frame(readRDS(file = glue("{dir_project}/2_pipeline/5_gateMarkers_cytokines/out/boolean_cytokines.rds")))

```

```{r}

sce_CD4 = q_SCEAddBoolMatrix(sce_CD4,boolean_matrix_cytokines[as.integer(sce_CD4$Original_cellID),])
sce_CD8 = q_SCEAddBoolMatrix(sce_CD8,boolean_matrix_cytokines[as.integer(sce_CD8$Original_cellID),])

```


### 2.2) Find combinations of cytokines

```{r}

boolean_result_CD4 = q_SCEBooleanCombinations(sce = sce_CD4,
                                          colnames_combine = c("IFN_pos","TNFa_pos","IL2_pos","CD107a_pos"))

boolean_result_CD8 = q_SCEBooleanCombinations(sce = sce_CD8,
                                          colnames_combine = c("IFN_pos","TNFa_pos","IL2_pos","CD107a_pos"))


```

```{r}

sce_CD4 = boolean_result_CD4$sce
sce_CD8 = boolean_result_CD8$sce

```

### 2.3) Subtract background

```{r}

df_colDataSCE_CD4 = data.frame(colData(sce_CD4))
df_colDataSCE_CD8 = data.frame(colData(sce_CD8))

df_colDataSCE = rbind(df_colDataSCE_CD4,
                      df_colDataSCE_CD8)

table(df_colDataSCE_CD8$sample_id,df_colDataSCE_CD8$Naive_memory)
table(df_colDataSCE$sample_id,df_colDataSCE$Naive_memory)

df_colDataSCE_CD4CD8M = df_colDataSCE %>%
  dplyr::filter(Naive_memory %in% c("CD4 Memory","CD8 Memory"))%>%
  select(Condition, Trial,PID,Visit, Stimulation,Naive_memory, Polyfunctionality_individual)

df_minusBackground_CD4CD8M = q_PsuedobulkPercPosMarker(df_colDataSCE_CD4CD8M, polyfunctionality = "Polyfunctionality_individual", background_col = "Stimulation", background_value = "Neg")

```


### 2.4) Plot results

```{r}
PIDs <- c("107","142","9214K","104","112","110", "113", "120","209")

df_preATI <- df_minusBackground_CD4CD8M %>%
  select(Condition,PID,Visit,Stimulation,Naive_memory,Polyfunctionality_individual,n,Total_cells)%>%
  rename(n_cells_pos = n,
         n_cells_tot = Total_cells)%>%
  merge(df_Timepoint%>%select(PID,Visit,weeks_from_preATI),
        by = c("PID","Visit"),
        all.x = T,
        all.y = F)%>%
  mutate(weeks_from_preATI = factor(weeks_from_preATI),
         PID = factor(PID, levels = PIDs),
         MarkerCount = str_count(Polyfunctionality_individual,"[+]"),
         Functionality = ifelse(MarkerCount>1,"Polyfunctional","Monofunctional"),
         Functionality2 = case_when(MarkerCount == 1 ~ "1",
                                    MarkerCount == 2 ~ "2",
                                    MarkerCount > 2 ~ ">2"),
         Functionality2 = factor(Functionality2, levels = c("1","2",">2")))%>%
  dplyr::filter(weeks_from_preATI==0)%>%
  select(-c(Visit,weeks_from_preATI))

# sum n positive cells per polyfunctionality individual (Functionality)
df_preATI_pooled <- df_preATI %>%
  group_by(PID,Condition,Naive_memory,Functionality,Stimulation)%>%
  summarise(n_cells_pos_sum = sum(n_cells_pos),
            n_cells_tot = n_cells_tot)%>% 
  unique()

# Calculate frequency 
df_freq <- df_preATI_pooled %>%
  group_by(PID, Condition, Naive_memory, Stimulation,
           n_cells_pos_sum, n_cells_tot) %>%
  mutate(
    Percentage = (n_cells_pos_sum / n_cells_tot) * 100
  ) %>%
  ungroup() 

# Subtrack background
df_freq_background_subtract <- df_freq %>% 
  select(-c(n_cells_pos_sum,n_cells_tot))%>%
  group_by(PID,Condition,Naive_memory,Functionality) %>%
  mutate(Percentage_backgroundSubtracted = pmax(Percentage - Percentage[Stimulation == "Neg"],0))
all(df_freq_background_subtract[df_freq_background_subtract$Stimulation=="Neg","Percentage_backgroundSubtracted"]==0)

```

```{r}

my_comparisons <- list(c("ART controls", "Post-intervention controllers"))

```


#### Figure 4A

```{r}
df_freq_background_subtract2 <- df_freq_background_subtract %>%
  dplyr::filter(Stimulation != "Env")%>% # Remove Env stim since many NCs don't have this 
  mutate(Stimulation_pooled = case_when(Stimulation == "Neg" ~ "Neg",
                                        TRUE ~ "Peptide"))%>%
  select(-Stimulation,Percentage)%>%
  group_by(PID,Condition,Naive_memory,Functionality,Stimulation_pooled)%>%
  summarise(Percentage_backgroundSubtracted_sum = sum(Percentage_backgroundSubtracted))

df_final <- df_freq_background_subtract2 %>%
  dplyr::filter(Stimulation_pooled!="Neg")
```

```{r}
p1<-df_final %>%
  mutate(Naive_memory = factor(Naive_memory, levels = c("CD8 Memory","CD4 Memory")),
         PID = case_when(PID == "9214K"~"9254",
                         PID == "113"~"702",
                         TRUE ~ PID),
         PID = factor(PID, levels = c("107","142","9254","104","112","110", "120","209", "702")),
         Condition = factor(Condition,levels = c("NC","PIC"),labels = c("ART controls","Post-intervention controllers")),
         Stimulation = "Gag+Pol+Nef")%>%
  ggplot(aes(x = Condition,
             y = Percentage_backgroundSubtracted_sum,
             fill = Condition))+
  geom_boxplot(inherit.aes = TRUE,outliers = F)+
  geom_point(aes(shape = PID),
             size = 2)+
  facet_nested(Naive_memory~ Functionality+Stimulation,
             scales = "free",
             switch = "y",
             strip = strip_nested(background_y = list(element_rect(fill = "grey90",  colour = "black")),
                                      background_x = list(element_part_rect(side = "b", colour = "black", fill = NA),                   
                                                          element_rect(colour = NA, fill = NA)),
                                      by_layer_y = T,
                                      by_layer_x = T))+
  ylab("Percentage of population (%)")+
  scale_shape_manual(values = c(19,17,15,0,1,2,3,6,8))+
  scale_fill_manual(values = c("ART controls" = "#C00000",
                                "Post-intervention controllers" = "#008F00"))+
  ggpubr::stat_compare_means(comparisons = my_comparisons,
                     method = "wilcox.test",
                     label = "p.adj",              # show adjusted p-value
                     p.adjust.method = "bonferroni",
                     size = 3)+
    theme(axis.text.x = element_blank(),
        legend.title=element_text(size = 14),
        legend.text = element_text(size = 10),
        legend.key=element_blank(),
          plot.title = element_text(hjust = 0.5),
          axis.text.y=element_text(size = 14),
          strip.text.x = element_text(size = 12),
          strip.text.y = element_text(size = 12),
          axis.title=element_text(size = 18, face="bold"),
          panel.grid.major.x = element_blank(),
          panel.grid.minor.x = element_blank(),
          panel.background = element_rect(fill = 'grey95'),
          panel.spacing.y = unit(1, "lines"))


```

#### Figure 4B

```{r}

df_final <- df_freq_background_subtract %>%
  dplyr::filter(Stimulation!="Neg")

```

```{r}

p2<-df_final %>%
  mutate(Naive_memory = factor(Naive_memory, levels = c("CD8 Memory","CD4 Memory")),
         Stimulation = factor(Stimulation, levels = c("Gag","Pol","Nef","Env")),
         PID = case_when(PID == "9214K"~"9254",
                         PID == "113"~"702",
                         TRUE ~ PID),
         PID = factor(PID, levels = c("107","142","9254","104","112","110", "120","209", "702")),
         Condition = factor(Condition,levels = c("NC","PIC"),labels = c("ART controls","Post-intervention controllers")))%>%
  ggplot(aes(x = Condition,
             y = Percentage_backgroundSubtracted,
             fill = Condition))+
  geom_boxplot(inherit.aes = TRUE,outliers = F)+
  geom_point(aes(shape = PID),
             size = 2)+
  facet_nested(Naive_memory~ Functionality+Stimulation,
             scales = "free",
             switch = "y",
             strip = strip_nested(background_y = list(element_rect(fill = "grey90",  colour = "black")),
                                      background_x = list(element_part_rect(side = "b", colour = "black", fill = NA),                   
                                                          element_rect(colour = NA, fill = NA)),
                                      by_layer_y = T,
                                      by_layer_x = T))+
  ylab("Percentage of population (%)")+
  scale_shape_manual(values = c(19,17,15,0,1,2,3,6,8))+
  scale_fill_manual(values = c("Post-intervention controllers" = "#008F00","ART controls" = "#C00000"))+
  ggpubr::stat_compare_means(comparisons = my_comparisons,
                     method = "wilcox.test",
                     label = "p.adj",              # show adjusted p-value
                     p.adjust.method = "bonferroni",
                     size = 3)+
  theme(axis.text.x = element_blank(),
        legend.title=element_text(size = 14),
        legend.text = element_text(size = 10),
        legend.key=element_blank(),
          plot.title = element_text(hjust = 0.5),
          axis.text.y=element_text(size = 14),
          strip.text.x = element_text(size = 12),
          strip.text.y = element_text(size = 12),
          axis.title=element_text(size = 18, face="bold"),
          panel.grid.major.x = element_blank(),
          panel.grid.minor.x = element_blank(),
          panel.background = element_rect(fill = 'grey95'),
          panel.spacing.y = unit(1, "lines"))

```

#### Figure 4C

```{r}
df_plot <- df_minusBackground_CD4CD8M %>%
  dplyr::mutate(Naive_memory = factor(Naive_memory, levels = c("CD8 Memory","CD4 Memory")),
                percent_pos_minusBackground = Percentage_backgroundSubtracted,
                Condition = str_replace(Condition, "PTC", "PIC"),
                Condition = factor(Condition, levels = c("PIC","NC")))


library(ggtext)

#add NA data for PID 112 preATI Env (missing)
NA_data1 = df_plot%>%dplyr::filter(PID == "112",
                        Visit == "13",
                        Stimulation == "Nef")%>%
  mutate(percent_pos_minusBackground = NA,
         Stimulation = "Env")

# add NA data for PID 107 week 12 Env (missing)
NA_data2 = df_plot%>%dplyr::filter(PID == "107",
                        Visit == "24",
                        Stimulation == "Nef")%>%
  mutate(percent_pos_minusBackground = NA,
         Stimulation = "Env")

df_plot = rbind(df_plot,NA_data1,NA_data2)


polyfunc_highlight = df_plot[,c("Polyfunctionality_individual","Percentage_backgroundSubtracted")] %>% 
  group_by(Polyfunctionality_individual) %>% 
  summarize(Pol_sum = sum(Percentage_backgroundSubtracted))%>%
  slice_max(Pol_sum, n = 8)%>%
  pull(Polyfunctionality_individual)


df_plot = df_plot %>% 
  mutate(fill_value = case_when(Polyfunctionality_individual %in% polyfunc_highlight ~ paste0("**",Polyfunctionality_individual,"**"),
                               TRUE ~ Polyfunctionality_individual
        ))

df_plot$Polyfunctionality_individual = factor(df_plot$Polyfunctionality_individual, levels = .findUniqueMarkerCombinations(c("IFN", "TNFa","IL2","CD107a"), full_names = F))
levels_legend_label = unique(df_plot[order(df_plot$Polyfunctionality_individual),"fill_value",drop=T])
df_plot$fill_value = factor(df_plot$fill_value, levels = levels_legend_label)

df_Timepoint <- read.csv(glue("{dir_store}/df_timepoint.csv"),row.names = 1)

df_plot = df_plot%>%
  merge(df_Timepoint%>%select(PID,Visit,xlab_manual),
        by = c("PID","Visit"),
        all.x = T,
        all.y = F)%>%
  mutate(xlab_manual = factor(xlab_manual),
         PID = case_when(PID == "9214K"~"9254",
                         PID == "113"~"702",
                         TRUE ~ PID),
         PID = factor(PID, levels = c("107","142","9254","104","112","110", "120","209", "702")),
         Condition = factor(Condition, levels = c("PIC","NC"),labels = c("Post-intervention controllers","ART controls")))%>%
  dplyr::filter(!is.na(xlab_manual))


# Desired y-limits per facet
limits <- data.frame(
  Naive_memory = c("CD8 Memory", "CD4 Memory"),
  ymin = c(0, 0),
  ymax = c(1.5, .55)
)
df_plot2 <- df_plot %>%
  left_join(limits, by = "Naive_memory") %>%
  mutate(percent_pos_minusBackground = pmin(pmax(percent_pos_minusBackground, ymin), ymax))
df_plot2$Naive_memory = factor(df_plot2$Naive_memory, levels = c("CD8 Memory","CD4 Memory"))

library(ggtext)
panel_spacing_single = unit(c(rep(c(rep(.5,4-1),2),3-1),rep(.5,4-1)),"lines")
p3 = q_plotPeptideSpecificResponse_NC_PTC(dataframe=df_plot2%>%
                                           dplyr::filter(!(PID == "142" & xlab_manual %in% c("135","140")),
                                                         Condition != "ART controls"), 
                                     fill_value = "fill_value",
                                     x_value = "xlab_manual",
                                     facet_y = "Naive_memory",
                                     subtract_background = T,
                                     legend_title = "Marker combination")+
  theme(legend.text = element_markdown(size=10),
        panel.spacing.x = panel_spacing_single)+
  xlab("ATI weeks")+
  geom_blank(aes(y = ymin)) +  # Forces the lower limit
  geom_blank(aes(y = ymax))   # Forces the upper limit

```

#### Figure 4

```{r}
design = "
AABBB
CCCCC
"

p1+p2+p3+plot_layout(guides = "collect",design = design,axis_titles = "collect_y")+plot_annotation(tag_levels = "A")&theme(plot.tag = element_text(face = 'bold',size = 20))
ggsave(glue("{dir_out}/Fig4.png"),width = 20, height = 15,units ="in")
ggsave(glue("{dir_out}/Fig4.tiff"),width = 20, height = 15,units ="in")
ggsave(glue("{dir_out}/Fig4.svg"),width = 20, height = 15,units ="in")
```



#### Extended data Fig. 3A

```{r}
panel_spacing_single = unit(c(rep(c(rep(.5,3),2),2),rep(c(rep(.5,3-1),2),4-1),rep(.5,3-1)),"lines")
p3_NCs = q_plotPeptideSpecificResponse_NC_PTC(dataframe=df_plot2%>%
                                           dplyr::filter(!(PID == "142" & xlab_manual %in% c("135","140")),
                                                         Condition == "ART controls"), 
                                     fill_value = "fill_value",
                                     x_value = "xlab_manual",
                                     facet_y = "Naive_memory",
                                     subtract_background = T,
                                     legend_title = "Marker combination")+
  theme(legend.text = element_markdown(size=10),
        panel.spacing.x = panel_spacing_single)+
  xlab("ATI weeks")+
  geom_blank(aes(y = ymin)) +  # Forces the lower limit
  geom_blank(aes(y = ymax))   # Forces the upper limit

p3_NCs
ggsave(glue("{dir_out}/ExtDataFig3A.png"),width = 15, height = 8, units = "in")
ggsave(glue("{dir_out}/ExtDataFig3A.tiff"),width = 15, height = 8, units = "in")
ggsave(glue("{dir_out}/ExtDataFig3A.svg"),width = 15, height = 8, units = "in")

```

#### Figure 7E

```{r}
q_plotPeptideSpecificResponse_NC_PTC(dataframe=df_plot2%>%dplyr::filter(PID=="142"), 
                                     fill_value = "fill_value",
                                     x_value = "xlab_manual",
                                     facet_y = "Naive_memory",
                                     subtract_background = T,
                                     legend_title = "Marker combination")+
  theme(legend.text = element_markdown(size=10))+
  xlab("ATI weeks")+
  geom_blank(aes(y = ymin)) +  # Forces the lower limit
  geom_blank(aes(y = ymax))   # Forces the upper limit

ggsave(file = glue("{dir_out}/Fig7E.svg"), height  = 9, width = 12, units = "in")
ggsave(file = glue("{dir_out}/Fig7E.tiff"), height = 9, width = 12, units = "in")
ggsave(file = glue("{dir_out}/Fig7E.png"), height = 9, width = 12, units = "in")

```



## 3) Session info

```{r Session info, message = TRUE}

sessionInfo()

```
